Problem: No way to design serving HTTP API
Solution:

 - Include gleam_cowboy as a dependency

Since cowboy is an industry-grade HTTP server on top of an industry-grade
socket acceptor "Ranch", we would like to piggy-back it to achieve nice
scalability and fault tolerance.

Currently, using it from gleam is a tiny challenge, since the wrapper was out
of date. This is why, for the time being, we are using our own maintenance fork
of gleam_cowboy.